move path rewrites to separate class
