+++ src/IIIFPresentation/Repository/Paths/PathRewriteParser.cs
@@ -1,0 +1,131 @@
+﻿using System.Text.RegularExpressions;
+using Core.Web;
+using Microsoft.Extensions.Options;
+using Models.API.General;
+
+namespace Repository.Paths;
+
+public interface IPathRewriteParser
+{
+    public PathParts ParsePathWithRewrites(string host, string path, int customer);
+}
+
+public class PathRewriteParser(IOptions<TypedPathTemplateOptions> options, ILogger<PathRewriteParser> logger) : IPathRewriteParser
+{
+    private readonly TypedPathTemplateOptions settings = options.Value;
+    
+    private const char PathSeparator = '/';
+    
+    /// <summary>
+    /// Parses a full URI into  the required path segemnts, taking into account path rewrites
+    /// </summary>
+    /// <param name="host"></param>
+    /// <param name="path">The path to match agains</param>
+    /// <param name="customer">The customer ID to use in the case that the path doesn't have one</param>
+    public PathParts ParsePathWithRewrites(string host, string path, int customer)
+    {
+        // Always try and parse canonical first
+        var canonical = ParseCanonical(path);
+        if (canonical != null) return canonical;
+        
+        // Not canonical - try and match to a path...
+        var templates = settings.GetPathTemplatesForHost(host);
+        var replacementRegex = new Regex("^{(.+)}$");
+        
+	    // First split the path into it's individual segments
+        var pathSplit = path.Split(PathSeparator,
+            StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
+		    
+	    foreach (var template in templates)
+	    {
+		    // Split template into chunks
+            var templateSplit = template.Value.Split(PathSeparator,
+                StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
+
+            // work out if the template is a FQDN and remove the host if it is
+            if (templateSplit.First().Contains("http")) templateSplit = templateSplit.Skip(2).ToArray();
+		    
+		    // Check lengths are same, if not don't compare, or it's possible to be just the host value
+            if (pathSplit.Length != templateSplit.Length &&
+                template.Key != PresentationResourceType.ResourcePublic) continue;
+
+            try
+            {
+                var (customerId, resourceId) =
+                    MatchValuesInTemplate(pathSplit, templateSplit, replacementRegex, customer);
+                if (resourceId != null)
+                {
+                    return new PathParts(customerId, resourceId,
+                        template.Key != PresentationResourceType.ResourcePublic);
+                }
+            }
+            catch (Exception e)
+            {
+                logger.LogError(e, "Error while parsing path {Path}", path);
+            }
+        }
+	    
+	    return new PathParts(null, null, false);
+    }
+
+    private static (int customerId, string? resourceId) MatchValuesInTemplate(string[] pathSplit, string[] templateSplit, 
+        Regex replacementRegex, int customer)
+    {
+        int? customerIdFromPath = null;
+        string? resourceId = null;
+
+        for (var i = 0; i < pathSplit.Length; i++)
+        {
+            var templatePart = templateSplit[i];
+            var valuePart = pathSplit[i];
+			    
+            // Check if this is a replacement value - if so get value from provided path 
+            if (replacementRegex.IsMatch(templatePart))
+            {
+                // This is a template - get the value of it from the path value
+                var capturedValue = replacementRegex.Match(templatePart).Groups[1].Value;
+                if (capturedValue == TypedPathTemplateOptions.SupportedTemplateOptions.CustomerId)
+                {
+                    customerIdFromPath = int.Parse(valuePart);
+                }
+                else if (capturedValue == TypedPathTemplateOptions.SupportedTemplateOptions.ResourceId) resourceId = valuePart;
+                else if (capturedValue == TypedPathTemplateOptions.SupportedTemplateOptions.HierarchyPath && 
+                         !SpecConstants.ProhibitedSlugs.Contains(valuePart))
+                {
+                    // everything in the path after hierarchy goes into the path
+                    resourceId = string.Join(PathSeparator, pathSplit.Skip(i));
+                    break;
+                }
+            }
+            else if (templatePart != valuePart)
+            {
+                // if this isn't a path replacement and the values don't match, abort
+                break;
+            }
+        }
+
+        // if the length is 1 less, and the template split is hierarchical, it means the root collection
+        if (pathSplit.Length == templateSplit.Length - 1 &&
+            templateSplit.Contains($"{{{TypedPathTemplateOptions.SupportedTemplateOptions.HierarchyPath}}}") &&
+            resourceId == null)
+        {
+            resourceId = string.Empty;
+        }
+
+        return (customerIdFromPath ?? customer, resourceId);
+    }
+    
+    private static PathParts? ParseCanonical(string path)
+    {
+        var canonicalRegex = new Regex("^\\/?(\\d+)\\/(manifests|collections|canvases)\\/(.+)$");
+
+        if (!canonicalRegex.IsMatch(path)) return null;
+	
+        var match = canonicalRegex.Match(path);
+        var customer = int.Parse(match.Groups[1].Value);
+        
+        return new PathParts(customer, match.Groups[3].Value, true);
+    }
+}
+
+public record PathParts(int? Customer, string? Resource, bool Hierarchical);
